test/backend/requirements.txt
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
pytest-xdist==3.5.0
factory-boy==3.3.0
faker==20.1.0
requests==2.31.0
django-extensions==3.2.3